Fix tests and a typo
