+++ tests/test_cache.py
@@ -8,8 +8,6 @@ Specific test class:
    python -m unittest tests.cache.TestCache
 
 """
-from unittest import TestSuite
-
 from tests.context import yfinance as yf
 
 import unittest
@@ -46,47 +44,6 @@ class TestCache(unittest.TestCase):
         cache.store(tkr, tz1)
 
         self.assertTrue(os.path.exists(os.path.join(self.tempCacheDir.name, "tkr-tz.db")))
-
-
-class TestCacheNoPermission(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        if os.name == "nt":  # Windows
-            cls.cache_path = "C:\\Windows\\System32\\yf-cache"
-        else:  # Unix/Linux/MacOS
-            # Use a writable directory
-            cls.cache_path = "/yf-cache"
-        yf.set_tz_cache_location(cls.cache_path)
-
-    def test_tzCacheRootStore(self):
-        # Test that if cache path in read-only filesystem, no exception.
-        tkr = 'AMZN'
-        tz1 = "America/New_York"
-
-        # During attempt to store, will discover cannot write
-        yf.cache.get_tz_cache().store(tkr, tz1)
-
-        # Handling the store failure replaces cache with a dummy
-        cache = yf.cache.get_tz_cache()
-        self.assertTrue(cache.dummy)
-        cache.store(tkr, tz1)
-
-    def test_tzCacheRootLookup(self):
-        # Test that if cache path in read-only filesystem, no exception.
-        tkr = 'AMZN'
-        # During attempt to lookup, will discover cannot write
-        yf.cache.get_tz_cache().lookup(tkr)
-
-        # Handling the lookup failure replaces cache with a dummy
-        cache = yf.cache.get_tz_cache()
-        self.assertTrue(cache.dummy)
-        cache.lookup(tkr)
-
-def suite():
-    ts: TestSuite = unittest.TestSuite()
-    ts.addTest(TestCache('Test cache'))
-    ts.addTest(TestCacheNoPermission('Test cache no permission'))
-    return ts
 
 
 if __name__ == '__main__':
